+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Field `arguments` added to the `amqp_0_9` input and output. (@calini)
+- Field `avro.mapping` added to the `schema_registry_decode` processor to support converting custom avro types to standard avro types for legacy tooling. (@rockwotj)
 
 ## 4.46.0 - 2025-01-29